counterbalancing is an ORDER, not a pair of equal counts
Round 22 finding 3. The check compared first-position COUNTS, which is a noun
standing in for the adjective "ordered by parity": a GROUPED order -- base-first
for the first half, head-first for the second -- has equal counts, passed, and
was REPORTED as parity-counterbalanced. Worse than most label-versus-property
instances because the record made an affirmative claim about experimental
design that the data did not support, and grouped order is precisely the
arrangement in which a time-varying drift correlates with arm order, which is
what counterbalancing exists to prevent.

Each replicate's observed first arm is now checked against the declared rule
rather than against an aggregate. The rule is not restated: pair_order is the
one statement of it, and a second statement would be a second thing to drift.

That required moving pair_order into ab_common, because ab_input and
ab_driver_support already import each other. A shared rule belongs in the
shared module rather than in whichever consumer happened to write it first.

560 cases (floor 560).
